scripts/jupyter/requirements.txt
pytest
cassandra-driver
jupyter
numpy
